refactor(ai/oauth): split grab-bag files by concern; rename token store
The package had two oversized grab-bag files. Split them along concern
boundaries (pure in-package moves — no symbol renames, no behavior change):

- token.go       (new)  Token type + its methods, out of config.go
- session.go     (new)  SessionState/SessionStatus + session mgmt, out of manager.go
- devicecode.go  (new)  DeviceCode* types + device-code flow/poll, out of config.go/manager.go
- tokenstore.go  (was store.go) renamed for consistency with sessionstore.go / statestore.go

manager.go now holds a focused auth-code core (1188 -> 838 lines) and
config.go is config-only (426 -> 290). No public API change; oauth and
consumer tests pass.
